feat(dataTable): split code
