+++ packages/ui/src/utils/types.ts
@@ -1,0 +1,1 @@
+export type Primitive = string | number | boolean | bigint | null | symbol;
